#958 - KB instance not properly filtered when annotating
- Factor auto-complete field for KB items into a separate re-usable class
- Use this class in the ConceptFeatureEditor and on the KB page
- Fix the search on the KB page mainly by properly handling cases where withLabelMatchingExacltlyAnyOf or withLabelContainingAnyOf are called with an emtpy argument list
